pyproject.toml
[tool.poetry]
name = "geo-service"
version = "0.1.0"
description = ""
authors = ["JKearnsl <boyarshinovn@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.109.0"
uvicorn = "^0.27.0.post1"

# Other
aiohttp = "^3.9.3"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
